+++ core/src/test/java/dev/failsafe/issues/Issue311Test.java
@@ -7,6 +7,7 @@ import org.testng.annotations.Test;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executor;
+import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -25,6 +26,7 @@ public class Issue311Test {
       })
       .get();
     assertEquals(counter.get(), 2);
+    ((ExecutorService) executor).shutdownNow();
   }
 
   public void testNullCompletionStage() throws Throwable {